Move Cesium-specific message handling into separate module
